+++ src/repl_tooling/repl_client.cljs
@@ -1,5 +1,5 @@
 (ns repl-tooling.repl-client
-  (:require [cljs.core.async :refer [chan <! >!] :refer-macros [go-loop go] :as async]
+  (:require [cljs.core.async :refer [<! >!] :refer-macros [go-loop go] :as async]
             [repl-tooling.repl-client.protocols :as repl]
             [clojure.string :as str]))
 
@@ -17,8 +17,15 @@
     (swap! sessions assoc session-name socket)
     [in out]))
 
+; FIXME! Really!
+(defn socket2! [session-name host port]
+  (let [[in out socket] (repl/connect-socket2! host port)]
+
+    (swap! sessions assoc session-name socket)
+    [in out]))
+
 (defn integrate-repl [in out repl]
-  (let [n-in (chan)]
+  (let [n-in (async/chan)]
     (go-loop []
       (>! in (repl/cmd-to-send repl (str (<! n-in))))
       (recur))