+++ src/repl_tooling/repl_client/clj_helper.clj
@@ -1,0 +1,5 @@
+(ns repl-tooling.repl-client.clj-helper
+  (:require [clojure.java.io :as io]))
+
+(defmacro blob-contents []
+  (slurp (io/resource "unrepl.clj")))