+++ src/repl_tooling/repl_client/lumo.cljs
@@ -1,7 +1,7 @@
 (ns repl-tooling.repl-client.lumo
   (:require [repl-tooling.repl-client.protocols :as repl]
             [repl-tooling.repl-client :as client]
-            [cljs.core.async :as async]
+            [cljs.core.async :as async :refer-macros [go]]
             [cljs.reader :as reader]
             [clojure.string :as str]))
 
@@ -19,16 +19,19 @@
 (defrecord Lumo [pending-cmds]
   repl/Repl
   (cmd-to-send [_ command]
-    (let [id (gensym)
-          cmd (code-to-lumo id command)]
+    (let [[id cmd] (if (str/starts-with? command "[")
+                     (reader/read-string command)
+                     [(gensym) command])
+          command (code-to-lumo id cmd)]
       (swap! pending-cmds conj (str id))
-      cmd)))
+      command)))
 
 (defn- treat-output [pending-cmds out]
   (let [[_ match] (re-find #"^\s*\[(.+?) " out)]
     (if (@pending-cmds match)
       (let [[_ out result] (reader/read-string out)]
-        {:out out :result result})
+        (swap! pending-cmds disj match)
+        {:id match :out out :result result})
       {:out out})))
 
 (defn connect-socket! [session-name host port]
@@ -37,5 +40,7 @@
         repl (->Lumo pending-cmds)
         [in out] (client/integrate-repl in out repl)
         new-out (async/map #(treat-output pending-cmds %) [out])]
+    (async/put! in '(require 'lumo.repl))
+    (async/put! in "(set! lumo.repl/*pprint-results* false)")
 
     [in new-out]))